pyproject.toml
[project]
name = "chatbotwithalbert"
version = "0.1.0"
description = ""
authors = [
    {name = "Swann",email = "swann.bouviermuller@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit (>=1.41.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "fastapi (>=0.115.7,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "openai (>=1.60.2,<2.0.0)",
    "fpdf2 (>=2.8.2,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.3"

